added NTP check #18
Added a new command to check NTP synchronisation with configured peers.
This is pretty much a reimplementation of check_ntp_peer. Output format and
performance data should match to 99%. Also the behaviour of the warning and
critical threshold checking got matched to to check_ntp_peer.

To avoid adding a couple of extra command line options the thresholds can be
set with following options. They have to be comma separated.
* o -> offset
* s -> stratum
* j -> jitter
* t -> truechimers

example:
 -C ntp -w o=0.03,s=1,t=3,j=100 -c o=0.05,s=2,j=200,t=2

Fixes: #18